Create a GeoLocator Data Package

This function create a Data package for GeoLocator data. This function is a wrapper of create_package() with additional constrains following GeoLocatorDP. The arguments of the functions correspond to the standard Data Package properties.

Usage

create_gldp(
  title,
  contributors,
  schema = NULL,
  name = NULL,
  id = NULL,
  licenses = list(list(name = "CC-BY-4.0", title = "Creative Commons Attribution 4.0",
    path = "https://creativecommons.org/licenses/by/4.0/")),
  description = NULL,
  homepage = NULL,
  image = NULL,
  version = NULL,
  embargo = "1970-01-01",
  keywords = NULL,
  citation = NULL,
  grants = NULL,
  related_identifiers = NULL,
  references = NULL
)

Arguments

title

A string providing a title or one sentence description for this package. It should be plain text (no markup), capitalised like a title, NOT end in a period and less than 65 characters. See Geolocator DP and thw Data Package specification.

contributors

A list of contributors, where each contributor is a list with properties including at leas title but also optionally givenName, familyName, path, email, roles, and organization. See the Geolocator DP and the Data Package specification . ](https://r-pkgs.org/description.html#sec-description-authors-at-r)

schema

(optional) A URL to the JSON Table Schema that describes the data package. Defaults to the GeoLocator Data Package profile. See the Data Package specification .

name

(optional) A string providing a name for the package. Defaults to a slugified version of the title. See the Geolocator DP and the Data Package specification.

id

A globally unique identifier for the package, such as a UUID or DOI. Defaults to NULL. See the Geolocator DP and the Data Package specification.

licenses

A list of licenses under which the package is provided. Each license is a list with properties name and path and optionally title. Defaults is a CC-BY-4.0 license. See the Geolocator DP and the Data Package specification.

description

A markdown-formatted string describing the package. You can (and should!) use multiple sentences, but limited to a single paragraph. See the Geolocator DP and the Data Package specification.

homepage

(optional) A URL for the home on the web related to this data package. See the Geolocator DP and the Data Package specification.

image

(optional) A URL or path pointing to the location of an image representing the data package. See the Geolocator DP and the Data Package specification.

version

(optional) A version string identifying the version of the package, following Semantic Versioning. Defaults to "1.0.0". See Geolocator DP and the Data Package specification and Data Package Version recipe.

embargo

End date of the embargo. Default to "1970-01-01". See the Geolocator DP and the Data Package specification.

keywords

(optional) A list of keywords to assist users in searching for the package. See the Geolocator DP and the Data Package specification.

citation

(optional) A string providing a citation for the package. See the Geolocator DP and the Data Package specification.

grants

(optional) A list of grants that funded the creation of the package. See the Geolocator DP and the Data Package specification.

related_identifiers

(optional) A list of related identifiers for the package. Each related identifier is a list with properties relationType and relatedIdentifier. See the Geolocator DP and the Camtrap DP specification .

references

(optional) List of references related to the package (e.g. references cited in package.description). References preferably include a DOI. See the Geolocator DP and the Camtrap DP specification.

Value

A list containing the descriptor for the data package.

Details

In addition, to the metadata entered as argument,

You can find more information about how to enter these meta-data information in the R package description.