Generate observations from GeoPressureR parameters

This function converts a list of parameters into a single observations tibble according to the GeoLocator Data Package specification .

Usage

params2o(params)

Arguments

params

A list of GeoPressureR parameter objects. These parameters should have been generated during the GeoPressure worflow. See GeoPressureR::param_create() for more information.

Value

A tibble with columns tag_id, manufacturer, scientific_name, ring_number, model, and firwmare (if soi_settings is present).