Read a GeoLocator Data Package

This function reads a GeoLocator Data Package from a "datapackage.json", file that describes the Data Package metadata and its Data Resources. The function wraps the function frictionless::read_package and assigns the class "geolocatordp" to the datapackage read.

Usage

read_gldp(file = "datapackage.json", force_read = TRUE)

Arguments

file

A string specifying the path to the JSON file containing the GeoLocator Data Package metadata. Defaults to "datapackage.json". Can also be a url.

force_read

Logical to force the reading of the data from path/url to memory.

Value

A GeoLocator Data Package object created from a file/url

Examples

# Read a datapackage.json file
pkg <- read_gldp("https://zenodo.org/records/14099115/files/datapackage.json")

pkg
#> 
#> ── A GeoLocator Data Package (main) 
#> • title: "GeoLocator Data Package: South African Woodland Kingfisher"
#> • contributors:
#>   Raphaël Nussbaumer (rafnuss@gmail.com) (Researcher, ProjectLeader) -
#>   <https://orcid.org/0000-0002-8185-1020>
#>   Yann Rime (yann.rime@vogelwarte.ch) (Researcher, Contributor) -
#>   <https://orcid.org/0009-0005-7264-6753>
#> • embargo: 1970-01-01
#> • licenses:
#>   CC BY 4.0
#> • id: <https://doi.org/10.5281/zenodo.13829929>
#> • version: "1.1"
#> • relatedIdentifiers:
#>   IsPartOf <10.5281/zenodo.11207081>
#>   IsSupplementTo <>
#> • grants: "Swiss Ornithological Intitute"
#> • keywords: "Woodland Kingfisher", "intra-african", and "multi-sensor
#>   geolocator"
#> • created: 2024-11-12 12:58:55
#> • spatial:
#> • temporal: "2017-01-10" to "2019-09-28"
#> • taxonomic: "Halcyon senegalensis"
#> • numberTags:
#> 
#> ── 8 resources: 
#> • tags
#> • observations
#> • measurements
#> • twilights
#> • staps
#> • paths
#> • edges
#> • pressurepaths
#> Use `unclass()` to print the Geolocator Data Package as a list.