Get GeoLocator DP version

Extracts the version number used by a GeoLocator Data Package object. This version number indicates what version of the GeoLocator DP standard was used.

The version number is derived as follows:

The version attribute, if defined.

A version number contained in x$profile, which is expected to contain the URL to the used GeoLocator DP standard.

x$profile in its entirety (can be NULL).

Usage

version(x)

Arguments

x

A GeoLocator Data Package object, as returned by create_gldp(). Also works on a Frictionless Data Package, as returned by read_package().

Value

Character string with the GeoLocator DP version number (e.g. "v0.2").