Update a GeoLocator Data Package Metadata

This function updates the metadata of an existing GeoLocator Data Package by setting or updating several fields based on the data contained within the package. Specifically, it updates the created,temporal, taxonomic, and reference_location fields based on the data resources present in the package.

Usage

update_gldp(pkg)

Arguments

pkg

An object of class "geolocatordp" representing the GeoLocator Data Package to be updated.

Value

An updated GeoLocator Data Package object of class "geolocatordp" with modified metadata.

Details

The function performs the following updates:

Sets the created field to the current date and time in RFC3339 format.

Updates the temporal field with the date range of the datetime values from the measurements resource, if available.

Updates the taxonomic field with unique, non-missing scientific_name values from the tags resource, if available.

Sets the reference_location field to the median latitude and longitude values from the observations resource, if available.